--- javascript-modules-engine/src/client-javascript/index.ts
@@ -1,5 +1,0 @@
-import { initI18next } from "./i18next/index.js";
-import { hydrateReactComponents } from "./react/index.js";
-
-initI18next();
-hydrateReactComponents();
