+++ javascript-modules-engine/src/client/index.ts
@@ -1,0 +1,2 @@
+import "./i18next.js";
+import "./hydrate.js";
